seleccionarCiudad

+++ app/src/main/java/com/example/appmoviles_parcial/repositorio/modelos/Ciudad.kt
@@ -5,8 +5,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Ciudad(
     val name: String,
-    val lat: Float,
-    val lon: Float,
+    val lat: Double,
+    val lon: Double,
     val country: String,
     val state: String = ""
 )
